add  type as it seems to be used by some applications
